requirements.txt
# Python dependencies

fastapi
uvicorn[standard]
python-dotenv
mne
matplotlib
anthropic
